requirements.txt
pyTelegramBotAPI==4.22.1
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1